Исправили замечания 1)Вынесите инициализацию room в даггер: Room.databaseBuilder(applicationContext,AppDatabase::class.java,"room_fatabase").build() 2) Переназовите SingleActivity, SinglePresenter на что-то более отражающее суть экрана 3) Вот так лучше не делать @RequiresApi(Build.VERSION_CODES.O). Лучше повысить minApi 4) Вот так тоже лучше не делать it.id!!,it.user!!,it.webFormatURL!!. Восклицательные знаки имею ввиду
