+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.geekbrains.pictviewer">
+    package="com.glindor227.pictviewer">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -10,14 +10,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".App"
+        android:name="com.glindor227.pictviewer.App"
         android:theme="@style/AppTheme.NoActionBar"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".view.single.SingleActivity"
+            android:name="com.glindor227.pictviewer.view.image.ImageActivity"
             android:label="@string/title_activity_simple"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".view.main.MainActivity">
+        <activity android:name="com.glindor227.pictviewer.view.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
