--- app/src/main/java/com/geekbrains/pictviewer/model/MainNote.kt
@@ -1,3 +1,0 @@
-package com.geekbrains.pictviewer.model
-
-data class MainNote(val id:Int, val name:String = "", val urlAvatar:String)
