--- app/src/main/java/com/geekbrains/pictviewer/model/ormStorage/AppDatabase.kt
@@ -1,11 +1,0 @@
-package com.geekbrains.pictviewer.model.ormStorage
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import com.geekbrains.pictviewer.model.ormStorage.dao.ImageUrl
-import com.geekbrains.pictviewer.model.ormStorage.dao.ImageUrlDAO
-
-@Database(entities = [ImageUrl::class],version = 1,exportSchema = false)
-abstract class  AppDatabase:RoomDatabase() {
-    abstract fun getDAO(): ImageUrlDAO
-}
