--- app/src/main/java/com/geekbrains/pictviewer/model/ormStorage/dao/ImageUrl.kt
@@ -1,7 +1,0 @@
-package com.geekbrains.pictviewer.model.ormStorage.dao
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "table_url")
-class ImageUrl(@PrimaryKey var id:Int = 0,var name:String, var url:String)
