--- app/src/main/java/com/geekbrains/pictviewer/model/ormStorage/repository/UrlRepository.kt
@@ -1,7 +1,0 @@
-package com.geekbrains.pictviewer.model.ormStorage.repository
-
-interface UrlRepository {
-    fun getUrls(urlCallback: UrlRepositoryImpl.UrlListRepositoryCallback)
-    fun addUrl(id:Int,name:String,url:String)
-    fun clear()
-}
