--- app/src/main/java/com/geekbrains/pictviewer/model/retrofit/ImagesRepository.kt
@@ -1,6 +1,0 @@
-package com.geekbrains.pictviewer.model.retrofit
-
-interface ImagesRepository {
-    fun getImages(query:String,repositoryCallback: ImagesRepositoryImpl.ImagesListRepositoryCallback)
-    fun getLargeImage(repositoryCallback: ImagesRepositoryImpl.ImageLargeRepositoryCallback, id :Int)
-}
