--- app/src/main/java/com/geekbrains/pictviewer/model/retrofit/pixabay/PixabayApi.kt
@@ -1,10 +1,0 @@
-package com.geekbrains.pictviewer.model.retrofit.pixabay
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.QueryMap
-
-interface PixabayApi {
-    @GET("api")
-    fun getImagesApi(@QueryMap parameter:Map<String, String>): Call<PixabayResponse>
-}
