--- app/src/main/java/com/geekbrains/pictviewer/view/main/Callback.kt
@@ -1,5 +1,0 @@
-package com.geekbrains.pictviewer.view.main
-
-interface Callback {
-    fun onItemClicked(id: Int)
-}
