--- app/src/main/java/com/geekbrains/pictviewer/view/single/SingleView.kt
@@ -1,5 +1,0 @@
-package com.geekbrains.pictviewer.view.single
-
-interface SingleView {
-    fun onImageComplete(imgUrl:String?)
-}
