+++ app/src/main/java/com/glindor227/pictviewer/model/MainNote.kt
@@ -1,0 +1,3 @@
+package com.glindor227.pictviewer.model
+
+data class MainNote(val id:Int?, val name:String? = "", val urlAvatar:String?)
