+++ app/src/main/java/com/glindor227/pictviewer/model/ormStorage/repository/UrlRepository.kt
@@ -1,0 +1,7 @@
+package com.glindor227.pictviewer.model.ormStorage.repository
+
+interface UrlRepository {
+    fun getUrls(urlCallback: UrlRepositoryImpl.UrlListRepositoryCallback)
+    fun clear()
+    fun addUrl(id: Int, name: String?, url: String?)
+}
