+++ app/src/main/java/com/glindor227/pictviewer/view/image/ImageView.kt
@@ -1,0 +1,5 @@
+package com.glindor227.pictviewer.view.image
+
+interface ImageView {
+    fun onImageComplete(imgUrl:String?)
+}
