+++ app/src/main/java/com/glindor227/pictviewer/view/main/Callback.kt
@@ -1,0 +1,5 @@
+package com.glindor227.pictviewer.view.main
+
+interface Callback {
+    fun onItemClicked(id: Int)
+}
